Cellula_3week_Mohamed_Ramadan/personal_branding_rag/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
python-dotenv
tiktoken